Feat-contract-termination-freelancer-connection-one-touch-modal #70 - 계약이 끝난 프리랜서 -> 자세히 보기 -> 프로젝트 다시 제안하기 -> 원터치 모달 연결
